pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-support-triage"
version = "0.1.0"
description = "A real-world customer support triage environment for OpenEnv"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openenv-core[core]>=0.2.3",
    "openai>=2.7.2",
    "pydantic>=2.8.2",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-cov>=5.0.0",
]

[project.scripts]
server = "server.app:main"
baseline = "support_triage_env.baseline.run_baseline:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["support_triage_env*", "server*"]

[tool.pytest.ini_options]
testpaths = ["tests"]

